+++ lib/actionizer/actions/index.rb
@@ -14,13 +14,13 @@ module Actionizer
             begin
               include self.available_extensions[obj.to_sym].constantize
             rescue NameError => e
-              raise "Could not resolve extension module '#{self.available_extensions[obj.to_sym]}' for key for #{obj.to_sym.inspect}. Check Actionizer/self.available_extensions[#{obj.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
+              raise ::Actionizer::ConfigurationError.new "Could not resolve extension module '#{self.available_extensions[obj.to_sym]}' for key for #{obj.to_sym.inspect}. Check Actionizer/self.available_extensions[#{obj.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
             end
           when String
             begin
               include obj.constantize
             rescue NameError => e
-              raise "Could not resolve extension module: #{obj}. Error: \n#{e.message}\n\n#{e.backtrace.join("\n")}"
+              raise ::Actionizer::ConfigurationError.new "Could not resolve extension module: #{obj}. Error: \n#{e.message}\n\n#{e.backtrace.join("\n")}"
             end
           else
             include obj