+++ lib/actionizer/configuration_error.rb
@@ -1,0 +1,4 @@
+module Actionizer
+  class ConfigurationError < StandardError
+  end
+end