+++ lib/actionizer/controller.rb
@@ -17,11 +17,11 @@ module Actionizer
       # You can define more or override in ::Actionizer.available_actions and/or self.available_actions.
       def include_actions(*args)
         args.each do |arg|
-          raise "#{arg.inspect} isn't defined in self.available_actions" unless self.available_actions[arg.to_sym]
+          raise ::Actionizer::ConfigurationError.new "#{arg.inspect} isn't defined in self.available_actions" unless self.available_actions[arg.to_sym]
           begin
             include self.available_actions[arg.to_sym].constantize
           rescue NameError => e
-            raise "Failed to resolve action module. Check Actionizer/self.available_actions[#{arg.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
+            raise ::Actionizer::ConfigurationError.new "Failed to resolve action module. Check Actionizer/self.available_actions[#{arg.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
           end
         end
       end
@@ -33,11 +33,11 @@ module Actionizer
       # You can define more or override in ::Actionizer.available_extensions and/or self.available_extensions.
       def include_extensions(*args)
         args.each do |arg|
-          raise "#{arg.inspect} isn't defined in self.available_extensions" unless self.available_extensions[arg.to_sym]
+          raise ::Actionizer::ConfigurationError.new "#{arg.inspect} isn't defined in self.available_extensions" unless self.available_extensions[arg.to_sym]
           begin
             include self.available_extensions[arg.to_sym].constantize
           rescue NameError => e
-            raise "Failed to resolve extension module. Check Actionizer/self.available_extensions[#{arg.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
+            raise ::Actionizer::ConfigurationError.new "Failed to resolve extension module. Check Actionizer/self.available_extensions[#{arg.to_sym.inspect}].constantize. Error: \n#{e.message}\n#{e.backtrace.join("\n")}"
           end
         end
       end